requirements.txt
requests
pandas
python-dotenv
matplotlib
selenium
bs4
webdriver_manager == 3.8.5
seaborn
Jpype1 == 1.4.1
konlpy
wordcloud
seaborn
scikit-learn
yellowbrick
spacy
numpy
transformers
torch
